Image_gen/requirements.txt
# requirements
streamlit
openai
python-dotenv
requests